mirror mode

+++ server/src/main/java/org/eclipse/openvsx/storage/StorageUtilService.java
@@ -9,24 +9,31 @@
  ********************************************************************************/
 package org.eclipse.openvsx.storage;
 
-import com.google.common.base.Strings;
-import com.google.common.collect.Maps;
+import static org.eclipse.openvsx.entities.FileResource.STORAGE_AZURE;
+import static org.eclipse.openvsx.entities.FileResource.STORAGE_DB;
+import static org.eclipse.openvsx.entities.FileResource.STORAGE_GOOGLE;
+
+import java.net.URI;
+import java.nio.file.Path;
+import java.util.AbstractMap;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
 import org.eclipse.openvsx.entities.ExtensionVersion;
 import org.eclipse.openvsx.entities.FileResource;
 import org.eclipse.openvsx.repositories.RepositoryService;
-import org.eclipse.openvsx.util.TimeUtil;
 import org.eclipse.openvsx.util.UrlUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Component;
 
-import java.net.URI;
-import java.nio.file.Path;
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static org.eclipse.openvsx.entities.FileResource.*;
+import com.google.common.base.Strings;
+import com.google.common.collect.Maps;
 
 /**
  * Provides utility around storing file resources and acts as a composite storage
@@ -49,6 +56,9 @@ public class StorageUtilService implements IStorageService {
 
     @Autowired
     DownloadCountService downloadCountService;
+
+    @Value("${ovsx.data.mirror.enabled:false}")
+    boolean readOnly;
 
     /** Determines which external storage service to use in case multiple services are configured. */
     @Value("${ovsx.storage.primary-service:}")
@@ -93,6 +103,10 @@ public class StorageUtilService implements IStorageService {
 
     @Override
     public void uploadFile(FileResource resource) {
+        if(readOnly) {
+            return;
+        }
+        
         var storageType = getActiveStorageType();
         switch (storageType) {
             case STORAGE_GOOGLE:
@@ -110,6 +124,10 @@ public class StorageUtilService implements IStorageService {
 
     @Override
     public void uploadFile(FileResource resource, Path filePath) {
+        if(readOnly) {
+            return;
+        }
+        
         var storageType = getActiveStorageType();
         switch (storageType) {
             case STORAGE_GOOGLE:
@@ -127,6 +145,10 @@ public class StorageUtilService implements IStorageService {
 
     @Override
     public void removeFile(FileResource resource) {
+        if(readOnly) {
+            return;
+        }
+
         switch (resource.getStorageType()) {
             case STORAGE_GOOGLE:
                 googleStorage.removeFile(resource);
@@ -178,7 +200,7 @@ public class StorageUtilService implements IStorageService {
     }
 
     public void increaseDownloadCount(FileResource resource) {
-        if(azureDownloadCountService.isEnabled()) {
+        if(readOnly || azureDownloadCountService.isEnabled()) {
             // don't count downloads twice
             return;
         }
